+++ src/bin/vip-dev-env-list.js
@@ -26,7 +26,7 @@ command( {
 	.argv( process.argv, async () => {
 		const lando = await bootstrapLando();
 		lando.events.constructor.prototype.setMaxListeners( 1024 );
-		await validateDependencies( lando, '' );
+		validateDependencies( lando );
 
 		const trackingInfo = { all: true };
 		await trackEvent( 'dev_env_list_command_execute', trackingInfo );